--- packages/apps/backend/tests/_factory/index.ts
@@ -1,2 +1,0 @@
-export * from './user.factory';
-export * from './admin-user.factory';
