+++ packages/apps/backend/tests/src/apis/utils/extract-token-from-http-headers.spec.ts
