+++ packages/apps/backend/tests/src/domain/aggregates/firebase-auth/firebase-auth.provider.spec.ts
