+++ packages/apps/backend/tests/src/utils/errors/transform-validation-error-into-detail.spec.ts
