+++ packages/apps/backend/tests/src/utils/is-empty-array.spec.ts
